Refactor graph operations into module
