+++ index.html
@@ -7,7 +7,7 @@
   <script src="https://unpkg.com/cytoscape/dist/cytoscape.min.js"></script>
   <script src="https://unpkg.com/dagre@0.8.5/dist/dagre.min.js"></script>
   <script src="https://unpkg.com/cytoscape-dagre@2.5.0/cytoscape-dagre.js"></script>
-  <script src="main.js?v0.2" defer></script>
+  <script type="module" src="dist/src/main.js"></script>
 
   <meta property="og:title" content="PIEFACE: RL Trace Visualizer">
   <meta property="og:description" content="Built at MIT CSAIL, PIEFACE helps debug RL agents trained for symbolic reasoning tasks.">
